requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-snowflake==5.2.1
apache-airflow-providers-microsoft-mssql==3.6.0
apache-airflow-providers-common-sql
azure-storage-blob
azure-identity
pymssql==2.2.11